requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-modeltranslation==0.19.7
idna==3.7
pillow==10.4.0
pyTelegramBotAPI==4.22.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
